Struct reth_net_common::bandwidth_meter::MeteredStream

pub struct MeteredStream<S> { /* private fields */ }

Wraps around a single stream that implements AsyncRead + AsyncWrite and meters the bandwidth through it

Implementations

impl<S> MeteredStream<S>

pub fn new(inner: S) -> Self

Creates a new MeteredStream wrapping around the provided stream, along with a new BandwidthMeter

pub fn new_with_meter(inner: S, meter: BandwidthMeter) -> Self

Creates a new MeteredStream wrapping around the provided stream, attaching the provided BandwidthMeter

pub fn get_bandwidth_meter(&self) -> &BandwidthMeter

Provides a reference to the BandwidthMeter attached to this MeteredStream

pub fn inner(&self) -> &S

Returns the wrapped stream